+++ models/contacts/index.js
@@ -1,0 +1,6 @@
+const { addContact } = require('./addContact')
+const { removeContact } = require('./removeContact')
+const { updateContact } = require('./updateContact')
+const { getContactById } = require('./getContactById')
+
+module.exports = { addContact, removeContact, updateContact, getContactById }
