+++ models/contacts/listContacts.js
@@ -1,0 +1,7 @@
+const { readContent } = require('../contactsActions.js')
+
+const listContacts = async () => {
+    return await readContent()
+}
+
+module.exports = { listContacts }
